+++ src/lantern/lib/metadata_library/models/record/elements/data_quality.py
@@ -6,6 +6,7 @@ import cattrs
 from lantern.lib.metadata_library.models.record.elements.common import Citation
 
 TDataQuality = TypeVar("TDataQuality", bound="DataQuality")
+TDomainConsistencies = TypeVar("TDomainConsistencies", bound="DomainConsistencies")
 
 
 @dataclass(kw_only=True)
@@ -40,6 +41,64 @@ class DomainConsistency:
     result: bool
 
 
+class DomainConsistencies(list[DomainConsistency]):
+    """
+    DomainConsistencies.
+
+    Wrapper around a list of DomainConsistency items with additional methods for filtering/managing items.
+
+    Schema definition: identifiers [1]
+    ISO element: gmd:report [2] (represents a single data quality element, multiple reports are allowed)
+
+    [1] https://github.com/antarctica/metadata-library/blob/v0.16.0/src/bas_metadata_library/schemas/dist/iso_19115_2_v4.json#L562
+    [2] https://www.datypic.com/sc/niem21/e-gmd_report-1.html
+    """
+
+    @classmethod
+    def structure(cls: type[TDomainConsistencies], value: list[dict]) -> "DomainConsistencies":
+        """
+        Parse domain consistency elements from plain types.
+
+        Returns a new class instance with parsed data. Intended to be used as a cattrs structure hook.
+        E.g. `converter.register_structure_hook(DomainConsistencies, lambda d, t: DomainConsistencies.structure(d))`
+
+        Structures input items into a list of DomainConsistency items via cattrs as a new instance of this class.
+
+        Example input: [{"specification": {"title": {"value": "x"}, "dates": {"creation": '2014-06-30'}}, "explanation": "x", "result": True}]
+        Example output: DomainConsistencies([DomainConsistency(specification=Citation(title="x", dates=Dates(creation=Date(date=date(2014, 6, 30)))), explanation="x", result=True)])
+        """
+        converter = cattrs.Converter()
+        converter.register_structure_hook(Citation, lambda d, t: Citation.structure(d))
+        return cls([converter.structure(domain, DomainConsistency) for domain in value])
+
+    def unstructure(self) -> list[dict]:
+        """
+        Convert to plain types.
+
+        Intended to be used as a cattrs unstructure hook.
+        E.g. `converter.register_unstructure_hook(DomainConsistencies, lambda d: d.unstructure())`
+
+        Example input: DomainConsistencies([DomainConsistency(specification=Citation(title="x", dates=Dates(creation=Date(date=date(2014, 6, 30)))), explanation="x", result=True)])
+        Example output: [{"specification": {"title": {"value": "x"}, "dates": {"creation": '2014-06-30'}}, "explanation": "x", "result": True}]
+        """
+        # noinspection PyUnresolvedReferences
+        converter = cattrs.Converter()
+        converter.register_unstructure_hook(Citation, lambda d: d.unstructure())
+        return [converter.unstructure(identifier) for identifier in self]
+
+    def filter(self, href: str) -> "DomainConsistencies":
+        """Filter domain consistency elements by specification href."""
+        return DomainConsistencies([domain for domain in self if domain.specification.href == href])
+
+    def ensure(self, domain: DomainConsistency) -> None:
+        """Add domain consistency elements without creating duplicates."""
+        if domain in self:
+            # skip exact match
+            return
+
+        self.append(domain)
+
+
 @dataclass(kw_only=True)
 class DataQuality:
     """
@@ -56,7 +115,7 @@ class DataQuality:
     """
 
     lineage: Lineage | None = None
-    domain_consistency: list[DomainConsistency] = field(default_factory=list)
+    domain_consistency: DomainConsistencies = field(default_factory=DomainConsistencies)
 
     @classmethod
     def structure(cls: type[TDataQuality], value: dict) -> "DataQuality":
@@ -76,7 +135,7 @@ class DataQuality:
             value["domain_consistency"][i] = dc
 
         converter = cattrs.Converter()
-        converter.register_structure_hook(Citation, lambda d, t: Citation.structure(d))
+        converter.register_structure_hook(DomainConsistencies, lambda d, t: DomainConsistencies.structure(d))
         return converter.structure(value, cls)
 
     def unstructure(self) -> dict:
@@ -87,7 +146,7 @@ class DataQuality:
         E.g. `converter.register_unstructure_hook(DataQuality, lambda d: d.unstructure())`
         """
         converter = cattrs.Converter()
-        converter.register_unstructure_hook(Citation, lambda d: d.unstructure())
+        converter.register_unstructure_hook(DomainConsistencies, lambda d: d.unstructure())
         value = converter.unstructure(self)
 
         # workaround v4 schema not allowing multiple contacts