+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 <!-- pyml disable no-duplicate-heading,no-duplicate-header -->
 ## [Unreleased]
 
+### Added
+
+* `DomainConsisties` container element for managing domain consistency data quality elements
+* `ensures()` util methods for contacts, identifiers and domain conformance elements to prevent duplicate items
+
 ## [0.5.0] - 2026-02-27 [!BREAKING]
 
 ### Changed [!BREAKING]